app/requirements/requirements.txt
alembic==1.4.2
bcrypt==3.1.7
blinker==1.4
cffi==1.14.0
click==7.1.2
cryptography==2.9.2
dnspython==1.16.0
email-validator==1.1.1
Flask==1.1.2
Flask-Caching==1.9.0
Flask-JWT-Extended==3.24.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
passlib==1.7.2
psycopg2-binary==2.8.4
pycparser==2.20
PyJWT==1.7.1
PyMySQL==0.9.3
pyOpenSSL==19.1.0
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
redis==3.5.3
six==1.15.0
SQLAlchemy==1.3.18
Werkzeug==1.0.1
WTForms==2.3.1

